TSLIB: Added support for multiple event reads from driver (at max 5).
This makes TSLIB more efficient. If "using_syn" is enabled
(default case), TSLIB expects to receive a series of events followed
by a SYN event in the read call to the driver. Once a SYN is
received, a flag (tsSampleready) is set to notify Eventhub that a
touch event (x,y,pressure,pen up/down,syn) is ready to be processed.
If a SYN is not received, the events received are simply stored in a
input_event array and the  control returns to EventHub.cpp.

Prevously, only one input event was being read in one read call in
input-raw.c. Also, TSLIB could not handle the case, where a SYN is
not received in a single ts_read call from EventHub.cpp. This may
occur at times when the TS driver is slow.

Modified the case, when event has type=EV_KEY,code=BTN_TOUCH
Here the "total" variable was being incremented, which is
inconsistent with "using_syn" approach.

"samp" variable was being incremented for "using_syn" case, this is not
needed as we only require one TS sample per ts_read call.

Change-Id: Ic584b13b5e943170ef3557cecffd4e754b4ecdb6
CRs-fixed: 283825